refactor: add relative path to contexts
