+++ apps/web/app/contacts/[creatorPubky]/page.tsx
@@ -3,12 +3,12 @@
 import Link from 'next/link';
 import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect, useState } from 'react';
+
 import { Button, Content, Icon, Typography } from '@social/ui-shared';
 import { CreatePost, Header } from '@/components';
-import { Contacts } from '../components';
 import Skeletons from '@/components/Skeletons';
-import { useClientContext } from '@/contexts/client';
-import { useFilterContext } from '@/contexts/filters';
+import { useClientContext, useFilterContext } from '@/contexts';
+import { Contacts } from '../components';
 import {
   IFollowingResponse,
   IFollowersResponse,
