+++ apps/web/app/contacts/components/_Me.tsx
@@ -1,7 +1,7 @@
 import { Icon, Typography } from '@social/ui-shared';
 import Image from 'next/image';
 import Link from 'next/link';
-import { useClientContext } from '@/contexts/client';
+import { useClientContext } from '@/contexts';
 import { Utils } from '@social/utils-shared';
 
 interface CountContacts {
