+++ apps/web/app/home/page.tsx
@@ -6,8 +6,7 @@ import { Content, Typography } from '@social/ui-shared';
 import * as Components from '@/components';
 import Skeletons from '@/components/Skeletons';
 import { Filter } from '@/components/Filter';
-import { useClientContext } from '@/contexts/client';
-import { useFilterContext } from '@/contexts/filters';
+import { useClientContext, useFilterContext } from '@/contexts';
 import { IPost, INewPost } from '@/types';
 
 export default function Index() {
