+++ apps/web/app/onboarding/sign-up/page.tsx
@@ -13,7 +13,7 @@ import {
   Icon,
   Typography,
 } from '@social/ui-shared';
-import { useClientContext } from '@/contexts/client';
+import { useClientContext } from '@/contexts';
 import { Modal } from '@/components/Modal';
 import { Utils } from '@social/utils-shared';
 import { Onboarding } from '../components';
