+++ apps/web/app/post/[pubky]/[postId]/components/_Replies.tsx
@@ -1,6 +1,6 @@
 import { Content, Icon, Button, Typography, SideCard } from '@social/ui-shared';
 import React, { useEffect, useState } from 'react';
-import { useClientContext } from '@/contexts/client';
+import { useClientContext } from '@/contexts';
 import { IReply } from '@/types';
 import { Post } from '@/components';
 import { Utils } from '@social/utils-shared';
