+++ apps/web/app/profile/[creatorPubky]/page.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 import { useEffect, useRef, useState } from 'react';
 import { Content, Typography } from '@social/ui-shared';
 import { CreatePost, Header, Post, PostsLayout } from '@/components';
-import { useClientContext } from '@/contexts/client';
+import { useClientContext } from '@/contexts';
 import { IPost, INewPost } from '@/types';
 import { Utils } from '@social/utils-shared';
 import Skeletons from '@/components/Skeletons';
