+++ apps/web/components/DropDown/_Status.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from 'react';
 import { DropDown as DropDownUI } from '@social/ui-shared';
 import { DropDown } from '..';
-import { useClientContext } from '@/contexts/client';
+import { useClientContext } from '@/contexts';
 import { TStatus } from '@/types';
 import { Utils } from '@social/utils-shared';
