+++ apps/web/components/Filter/_Reach.tsx
@@ -1,5 +1,5 @@
 import { Icon, SideCard } from '@social/ui-shared';
-import { useFilterContext } from '@/contexts/filters';
+import { useFilterContext } from '@/contexts';
 import { useEffect, useState } from 'react';
 import { TReach } from '@/types';
