+++ apps/web/components/Post/_Tags.tsx
@@ -9,7 +9,7 @@ import {
   PostUtil,
   Tooltip as TooltipUI,
 } from '@social/ui-shared';
-import { useClientContext } from '@/contexts/client';
+import { useClientContext } from '@/contexts';
 import { IPost, ITaggedPost } from '@/types';
 import { Utils } from '@social/utils-shared';
 import Tooltip from '../Tooltip';
