+++ apps/web/components/Tooltip/_Profile.tsx
@@ -10,7 +10,7 @@ import {
 } from '@social/ui-shared';
 import { Utils } from '@social/utils-shared';
 import { IPost, IFollowingResponse, IFollowersResponse } from '@/types';
-import { useClientContext } from '@/contexts/client';
+import { useClientContext } from '@/contexts';
 import { useRouter } from 'next/navigation';
 
 interface ProfileProps {
