+++ apps/web/components/_CreateQuickPost.tsx
@@ -7,12 +7,11 @@ import EmojiPicker, {
   EmojiStyle,
   Theme,
 } from 'emoji-picker-react';
-import { useClientContext } from '@/contexts/client';
+import { useClientContext, useAlertContext } from '@/contexts';
 import Image from 'next/image';
 import { INewPost } from '@/types';
 import Modal from './Modal';
 import { Utils } from '@social/utils-shared';
-import { useAlertContext } from '@/contexts/alerts';
 import LinkPreviewer from './_LinkPreviewer';
 
 export default function CreateQuickPost() {
