+++ apps/web/components/_Header.tsx
@@ -12,7 +12,7 @@ import {
   Menu,
 } from '@social/ui-shared';
 import { Modal } from './Modal';
-import { useClientContext } from '@/contexts/client';
+import { useClientContext } from '@/contexts';
 import { Utils } from '@social/utils-shared';
 
 interface HeaderProps {
