+++ apps/web/contexts/index.tsx
@@ -1,0 +1,12 @@
+import { useAlertContext, AlertWrapper } from './alerts';
+import { useClientContext, ClientWrapper } from './client';
+import { useFilterContext, FilterWrapper } from './filters';
+
+export {
+  useAlertContext,
+  useClientContext,
+  useFilterContext,
+  AlertWrapper,
+  ClientWrapper,
+  FilterWrapper,
+};
